Move `Serialized` and `bincode_config` a separate module
